+++ package/rhythmicon-app/components/RhythmEditor.vue
@@ -9,72 +9,123 @@ const rhythm = defineModel({ validator: r => r instanceof Rhythm })
 defineProps({ pulse: Number })
 
 const divisor = computed(() => rhythm.value.divisor())
-
-function duplicate() {
-  rhythm.value.push(...rhythm.value)
-}
-/* TOOD: if repeated
-function halve() { // TODO: if length >= 4
-  rhythm.value.splice(rhythm.value.length / 2, rhythm.value.length / 2)
-}
-*/
-
-function complement() {
-  rhythm.value.replace(...rhythm.value.map(x => x ? 0 : 1))
-}
-function append() {
-  rhythm.value.push(0)
-}
-function pop() {
-  rhythm.value.pop()
-}
 </script>
 
 <template>
   <div class="rhythm-editor">
-    <div>
-      <button class="action" :disabled="rhythm.empty()" @click="rhythm.rotate(-1)">
-        &lt;
-      </button>
-      <button class="action" :disabled="rhythm.empty()" @click="rhythm.rotateBeat(-1)">
-        ⋖
-      </button>
-      <button class="action" :disabled="rhythm[0] != 1" @click="rhythm.rotateBeat(1)">
-        ⋗
-      </button>
-      <button class="action" :disabled="rhythm.empty()" @click="rhythm.rotate(1)">
-        &gt;
-      </button>
-      <button class="action" @click="append">
-        +
-      </button>
-      <button class="action" :disabled="rhythm.length < 2" @click="pop">
-        -
-      </button>
-      <button class="action" @click="complement">
-        ⇅
-      </button>
-      <button class="action" @click="duplicate">
-        𝄎
-      </button>
-      <!--button class="action" :disabled="rhythm.length < 2 || rhythm.length % 2" @click="halve">
-        ½
-      </button-->
-      <button class="action" :disabled="divisor === 1" @click="rhythm.deflate()">
-        ÷{{ divisor > 1 ? divisor : "n" }}
-      </button>
-      <button class="action" :disabled="rhythm.length > MAX/2" @click="rhythm.inflate()">
-        ×2
-      </button>
-      <button class="action" :disabled="rhythm.length > MAX/3" @click="rhythm.inflate(3)">
-        ×3
-      </button>
-      <button class="action" :disabled="rhythm.length % 2 || rhythm.empty() || rhythm.length > 2*MAX/3" @click="rhythm.shuffle()">
-        ²=³
-      </button>
-      <button class="action" :disabled="!rhythm.isShuffle()" @click="rhythm.unshuffle()">
-        ³=²
-      </button>
-    </div>
+    <button
+      class="action" :disabled="rhythm.empty()" 
+      title="rotate one pulse left" @click="rhythm.rotate(-1)"
+    >
+      &lt;
+    </button>
+    <button
+      class="action" :disabled="rhythm.empty()" 
+      title="rotate one beat left" @click="rhythm.rotateBeats(rhythm[0] ? -1 : 0)"
+    >
+      ⋖
+    </button>
+    <button
+      class="action" :disabled="rhythm.empty()"
+      title="rotate one beat right" @click="rhythm.rotateBeats(1)"
+    >
+      ⋗
+    </button>
+    <button
+      class="action" :disabled="rhythm.empty()"
+      title="rotate one pulse right" @click="rhythm.rotate(1)"
+    >
+      &gt;
+    </button>
+    <button
+      class="action" title="add one pulse" 
+      @click="rhythm.push(0)"
+    >
+      +
+    </button>
+    <button
+      class="action" :disabled="rhythm.length < 2"
+      title="remove last pulse" @click="rhythm.pop()"
+    >
+      -
+    </button>
+    <button
+      class="action" title="switch beats and pauses"
+      @click="rhythm.complement()"
+    >
+      ⇅
+    </button>
+    <button
+      class="action" title="reverse rhythm"
+      @click="rhythm.reverse()"
+    >
+      ⇆
+    </button>
+    <button
+      class="action" title="repeat rhythm"
+      @click="rhythm.repeat()"
+    >
+      𝄎
+    </button>
+    <button
+      class="action" title="cut rhythm"
+      :disabled="rhythm.repetitions() == 1"
+      @click="rhythm.cut()"
+    >
+      𝄍
+    </button>    
+    <button
+      class="action" :disabled="divisor === 1"
+      title="deflate rhythm"
+      @click="rhythm.deflate()"
+    >
+      ÷{{ divisor > 1 ? divisor : "n" }}
+    </button>
+    <button
+      class="action" :disabled="rhythm.length > MAX/2"
+      title="inflate rhythm (double)" @click="rhythm.inflate()"
+    >
+      ×2
+    </button>
+    <button
+      class="action" :disabled="rhythm.length > MAX/3"
+      title="inflate rhythm (triple)" @click="rhythm.inflate(3)"
+    >
+      ×3
+    </button>
+    <button
+      class="action" :disabled="rhythm.length % 2 || rhythm.empty() || rhythm.length > 2*MAX/3"
+      title="shuffle rhythm" @click="rhythm.shuffle()"
+    >
+      ²=³
+    </button>
+    <button
+      class="action" :disabled="!rhythm.isShuffle()"
+      title="unshuffle rhythm" @click="rhythm.unshuffle()"
+    >
+      ³=²
+    </button>
   </div>
 </template>
+
+<style>
+.rhythm-editor {
+  position: fixed;
+  right: 0.2em;
+  top: 2.2em;
+  display: flex;
+  flex-direction: column;
+  border-right: none;
+  border-radius: 4px 0 0 4px;
+  padding: 0.25em;
+  gap: 0.25em;
+  z-index: 1000;
+}
+.rhythm-editor button {
+  font-weight: bold;
+}
+.rhythm-editor button:disabled {
+  font-weight: normal;
+}
+</style>
+