+++ package/rhythmicon-vue/CHANGES.md
@@ -1,6 +1,6 @@
 # 0.2.1
 
-- RhythCircle: show pulse index on hover
+- RhythCircle: show pulse index on hover, show symmetry
 - RhythmInput: allow more characters
 
 # 0.2.0